Camera Adjustments.
The camera script is now active when the game is not running, so changes in the inspector will
immediately change the camera in the scene according to its rules.

Yaw is now wrapped between -360 and 360, just to prevent it becoming a ridiculous number and
crashing the game, although I know that would never really happen.

Pitch is clamped between some values, so the player can't look too high or too low of an angle.