Adds movement scripts and test scene
The movement scripts are a state machine. If you want to add a
type of movement (such as grapple) you need to make a new script
that inherits from "MovementState", then implement the methods
you need. Add the state to those in "MovementController" and add
the script to the player object. You should not put code in
Update or FixedUpdate because it will be called even when the
movement state is not active.