Turret refinements.
The bullet is now a high visibility material. The turret raycast
layermask will now detect the player. The bullets are fired more
slowly so they are visible for more than 1 frame. The turret now
uses an array of spawn positions to determine where to originate
the bullet from and cycles between them. The bullet will despawn
after 1 second (currently set in turret).