Player camera now follows player rotation.
Basically if player is upside down, camera is upside down.
If player is sideways, camera is sideways.
This commit leaves much to be desired. The way this is implemented
the horizontal rotation is handled by cling state *throws up*.
Settings for horizontal sensitivity is on cling state component.
It is also very jerky, but I'm planning to do smoothing stuff soon.
I'm also planning to set up playerpref stuff so you will have local
settings for sensitivity, control inversion, etc.